Move baseUrl into constants folder
