+++ src/constants/baseUrl.ts
@@ -1,0 +1,4 @@
+// GraphQL endpoint
+const BASE_URL = process.env.BASE_URL || 'http://localhost:3000/graphql'
+
+export default BASE_URL
