+++ src/pages/_app.tsx
@@ -2,7 +2,7 @@ import { cacheExchange, fetchExchange } from 'urql'
 import { withUrqlClient } from 'next-urql'
 import { AppProps } from 'next/app'
 
-import { BASE_URL } from '@/util/constants'
+import BASE_URL from '@/constants/baseUrl'
 import '@/styles/globals.css'
 
 function App({ Component, pageProps }: AppProps) {
