--- src/util/constants.ts
@@ -1,3 +1,0 @@
-// GraphQL endpoint
-/* eslint-disable import/prefer-default-export */ // remove this once more constants are added
-export const BASE_URL = process.env.BASE_URL || 'http://localhost:3000/graphql'
